requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pyasn1==0.4.8
python-telegram-bot==13.6
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1
